Name changes for Adapter\ObjectRepository Storage\ObjectRepository Paginator\Adapter\Collection and Paginator\Adapter\Selectable
