+++ src/DoctrineModule/Paginator/Adapter/CollectionAdapter.php
@@ -19,41 +19,30 @@
 
 namespace DoctrineModule\Paginator\Adapter;
 
-use Doctrine\Common\Collections\Selectable as DoctrineSelectable;
-use Doctrine\Common\Collections\Criteria;
+use Doctrine\Common\Collections\Collection as DoctrineCollection;
 use Zend\Paginator\Adapter\AdapterInterface;
 
 /**
- * Provides a wrapper around a Selectable object
+ * Base module for Doctrine ORM.
  *
  * @license MIT
  * @link    http://www.doctrine-project.org/
  * @author  Michaël Gallego <mic.gallego@gmail.com>
  * @author  Marco Pivetta <ocramius@gmail.com>
  */
-class Selectable implements AdapterInterface
+class CollectionAdapter implements AdapterInterface
 {
     /**
-     * @var DoctrineSelectable
+     * @var DoctrineCollection
      */
-    protected $selectable;
+    protected $collection;
 
     /**
-     * @var \Doctrine\Common\Collections\Criteria
+     * @param DoctrineCollection $collection
      */
-    protected $criteria;
-
-    /**
-     * Create a paginator around a Selectable object. You can also provide an optional Criteria object with
-     * some predefined filters
-     *
-     * @param \Doctrine\Common\Collections\Selectable    $selectable
-     * @param \Doctrine\Common\Collections\Criteria|null $criteria
-     */
-    public function __construct(DoctrineSelectable $selectable, Criteria $criteria = null)
+    public function __construct(DoctrineCollection $collection)
     {
-        $this->selectable = $selectable;
-        $this->criteria   = $criteria ? clone $criteria : new Criteria();
+        $this->collection = $collection;
     }
 
     /**
@@ -61,9 +50,7 @@ class Selectable implements AdapterInterface
      */
     public function getItems($offset, $itemCountPerPage)
     {
-        $this->criteria->setFirstResult($offset)->setMaxResults($itemCountPerPage);
-
-        return $this->selectable->matching($this->criteria)->toArray();
+        return array_values($this->collection->slice($offset, $itemCountPerPage));
     }
 
     /**
@@ -71,11 +58,6 @@ class Selectable implements AdapterInterface
      */
     public function count()
     {
-        $criteria = clone $this->criteria;
-
-        $criteria->setFirstResult(null);
-        $criteria->setMaxResults(null);
-
-        return count($this->selectable->matching($criteria));
+        return count($this->collection);
     }
 }
