+++ tests/DoctrineModuleTest/Authentication/Storage/ObjectRepositoryStorageTest.php
@@ -20,7 +20,7 @@
 namespace DoctrineModuleTest\Authentication\Storage;
 
 use PHPUnit_Framework_TestCase as BaseTestCase;
-use DoctrineModule\Authentication\Storage\ObjectRepository as ObjectRepositoryStorage;
+use DoctrineModule\Authentication\Storage\ObjectRepositoryStorage;
 use DoctrineModuleTest\Authentication\Adapter\TestAsset\IdentityObject;
 use Zend\Authentication\Storage\NonPersistent as NonPersistentStorage;
