+++ tests/DoctrineModuleTest/Paginator/Adapter/SelectableAdapterTest.php
@@ -20,7 +20,7 @@
 namespace DoctrineModuleTest\Paginator\Adapter;
 
 use Doctrine\Common\Collections\Criteria;
-use DoctrineModule\Paginator\Adapter\Selectable as SelectableAdapter;
+use DoctrineModule\Paginator\Adapter\SelectableAdapter;
 use Doctrine\Common\Collections\ArrayCollection;
 use PHPUnit_Framework_TestCase;
